refactor(SLB-203): use operations to retrieve menus
