+++ packages/schema/src/schema.graphql
@@ -53,14 +53,17 @@ type NavigationItem {
 }
 
 interface Navigation {
+  locale: Locale!
   items: [NavigationItem]!
 }
 
 type MainNavigation implements Navigation @menu(menu_id: "main") {
+  locale: Locale! @resolveEntityLanguage
   items: [NavigationItem]! @lang @resolveMenuItems
 }
 
 type FooterNavigation implements Navigation @menu(menu_id: "footer") {
+  locale: Locale! @resolveEntityLanguage
   items: [NavigationItem]! @lang @resolveMenuItems
 }
 
@@ -164,6 +167,8 @@ type Query {
   previewDecapPage: Page
   previewDrupalPage(id: ID!, rid: ID, locale: String!): Page
     @fetchEntity(type: "node", id: "$id", rid: "$rid", language: "$locale")
+  mainNavigations: [MainNavigation] @gatsbyNodes(type: "MainNavigation")
+  footerNavigations: [FooterNavigation] @gatsbyNodes(type: "FooterNavigation")
   contentHub(query: String, pagination: PaginationInput!): ContentHubResult!
     @contentHub(query: "$query", pagination: "$pagination")
 }
